XDMAC: Moved errata transfer config to `transfer` module
